Migrated to Tanuki for scenario generation
